requirements.txt
basicSR
addict
future
lmdb
numpy>=2.2.6
opencv-python>=4.9.0
Pillow
pyyaml
requests
scikit-image
scipy
tb-nightly
torch>=1.7
torchvision
tqdm
yapf
pytesseract==0.3.13
pytz
realesrgan